+++ packages/wallet-stack/src/send/useSendFilterChips.ts
@@ -5,9 +5,19 @@ import { DynamicConfigs } from 'src/statsig/constants'
 import { StatsigDynamicConfigs } from 'src/statsig/types'
 import { TokenBalance } from 'src/tokens/slice'
 
-export default function useSendFilterChips(isMiniPayRecipient?: boolean): {
+export default function useSendFilterChips({
+  isMiniPayRecipient,
+  tokens,
+  defaultTokenIdOverride,
+  lastUsedTokenId,
+}: {
+  isMiniPayRecipient?: boolean
+  tokens: TokenBalance[]
+  defaultTokenIdOverride?: string
+  lastUsedTokenId?: string | null
+}): {
   filterChips: BooleanFilterChip<TokenBalance>[]
-  miniPayTokenIds: string[] | null
+  defaultToken: TokenBalance | undefined
 } {
   const { miniPayTokenIds: configTokenIds } = getDynamicConfigParams(
     DynamicConfigs[StatsigDynamicConfigs.SEND_CONFIG]
@@ -29,5 +39,15 @@ export default function useSendFilterChips(isMiniPayRecipient?: boolean): {
     [miniPayTokenIds]
   )
 
-  return { filterChips, miniPayTokenIds }
+  const defaultToken = useMemo(() => {
+    const eligibleTokens = miniPayTokenIds
+      ? tokens.filter((token) => miniPayTokenIds.includes(token.tokenId))
+      : tokens
+    const defaultToken = eligibleTokens.find((token) => token.tokenId === defaultTokenIdOverride)
+    const lastUsedToken = eligibleTokens.find((token) => token.tokenId === lastUsedTokenId)
+
+    return defaultToken ?? lastUsedToken ?? eligibleTokens[0]
+  }, [tokens, miniPayTokenIds, defaultTokenIdOverride, lastUsedTokenId])
+
+  return { filterChips, defaultToken }
 }
